requirements.txt
bs4
patch
py-cpuinfo
distro
iopath
pytest
pytest-benchmark
requests
tabulate
timm == 0.4.12
MonkeyType
psutil
pyyaml
numpy
